# config/openapi_templates/yaml/openapi_schema.yaml
# Copyright 2023-2026 Airbus, CS Group
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# -------------- CAUTON: this schema is defined both in rs-server and rs-dpr-service
# repos (/config/openapi_templates/openapi_schema.yaml) - if the current file
# is updated, don't forget to update the file from the other repository

openapi: 3.0.0
info:
  title: RS DPR service API
  version: 1.0.0
paths:
  /dpr/processes:
    get:
      summary: Get a list of processes
      responses:
        '200':
          description: Returns list of all available processes from config.
          content:
            application/json:
              schema:
                $ref: './processList.yaml'
  /dpr/processes/{process_id}:
    get:
      summary: Get information on a specific process
      parameters:
        - name: process_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns a description of a specific process
          content:
            application/json:
              schema:
                $ref: './process.yaml'
        '404':
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
            text/html:
              schema:
                type: string
  /dpr/processes/{process_id}/execution:
    post:
      summary: Execute a process
      parameters:
        - name: process_id
          in: path
          required: true
          schema:
            type: string
      # Here both the request and the response need to be validates according to yaml schemas
      requestBody:
        description: Check the body from the request used to execute the process
        content:
          application/json:
            schema:
              $ref: './execute.yaml'
        required: true

      responses:
        '200':
          description:  Result of synchronous execution
          content:
            application/json:
              schema:
                $ref: './statusInfo.yaml'
        '201':
          description: Started asynchronous execution. Created job.
          content:
            application/json:
              schema:
                $ref: './statusInfo.yaml'
        "404":
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
        "500":
          description: A server error occurred.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
  /dpr/jobs:
    get:
      summary: Returns the status of all jobs.
      responses:
        '200':
          description: Returns the status of all jobs.
          content:
            application/json:
              schema:
                $ref: './jobList.yaml'
        '404':
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
            text/html:
              schema:
                type: string
  /dpr/jobs/{job-id}:
    get:
      summary: Returns the status of a job.
      parameters:
        - name: job-id
          in: path
          description: local identifier of a job
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The status of a job.
          content:
            application/json:
              schema:
                $ref: 'statusInfo.yaml'
        '404':
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
            text/html:
              schema:
                type: string
        '500':
          description: A server error occurred.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
            text/html:
              schema:
                type: string
    delete:
      summary: Deletes a specific job.
      parameters:
        - name: job-id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the job to be deleted.
      responses:
        '200':
          description: Deletes a specific job.
          content:
            application/json:
              schema:
                $ref: './statusInfo.yaml'
        '404':
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
            text/html:
              schema:
                type: string
        '500':
          description: A server error occurred.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
            text/html:
              schema:
                type: string
  /dpr/jobs/{job-id}/results:
    get:
      summary: Returns the results of a given job
      parameters:
        - name: job-id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns the status of all jobs.
          content:
            application/json:
              schema:
                $ref: './results.yaml'
        '404':
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
            text/html:
              schema:
                type: string
        '500':
          description: A server error occurred.
          content:
            application/json:
              schema:
                $ref: './exception.yaml'
            text/html:
              schema:
                type: string